Clear flags when reversing prior to do other manual corrections
Firts clear flags of previous changes made by treenetproc, add the 'rev' flag and then do the other manual corrections.

This comes out from the need that sometimes we want to reverse a change made by treenetproc and, at the same item position, force a jump. With the previous code, the force jump got a NA value since there was a 'fill' flag when the forcejump function was executing (see second statement of the forcejump function). This was because removing the previous flags and setting the 'rev' flag was done after all the corrections are made.